perf: list dependency detail by distribution
